added responsiveness
added mediaqueries and played with flexbox and grid container. It's doens't work completely yet, because i need to change the width of the charts in js with:
function beforePrintHandler () {
    for (var id in Chart.instances) {
        Chart.instances[id].resize();
    }
}
Because i never worked with mediaqueries in js, it's taking longer than I would have wanted. But now the code isn't working, because the boxes of the charts get to big when the screen gets bigger.

Tommorow I will try to add this function and if it isn't working, I will put everything in a different grid-item and fix the problem.